+++ src/components/form/RegisterBusinessForm.jsx
@@ -23,7 +23,7 @@ import {
   REGISTER_BUSINESS_FORM_NAME,
 } from '../../constants';
 
-import createIntl from '../../util/createIntl';
+import createINTL from '../../util/createINTL';
 import encodeFormState from '../../util/encodeFormState';
 
 const FormContainer = styled.div`
@@ -94,7 +94,7 @@ class RegisterBusinessForm extends React.Component {
   }
 
   validatePageData() {
-    const i = createIntl(this.props.intl);
+    const i = createINTL(this.props.intl);
     const errors = {};
     let isValid = true;
 
@@ -298,7 +298,7 @@ class RegisterBusinessForm extends React.Component {
   }
 
   render() {
-    const i = createIntl(this.props.intl);
+    const i = createINTL(this.props.intl);
     const { locale } = this.props;
 
     return (