+++ src/templates/authors.jsx
@@ -14,7 +14,7 @@ import Rule from '../components/Rule';
 import SchemedImage from '../components/SchemedImage';
 import Theme from '../components/Theme';
 
-import createIntl from '../util/createIntl';
+import createINTL from '../util/createINTL';
 
 const Separator = styled.div`
   display: flex;
@@ -130,7 +130,7 @@ const TwitterImage = styled(SocialMediaImage)`
 `;
 
 function Page(props) {
-  const i = createIntl(useIntl());
+  const i = createINTL(useIntl());
 
   const { edges: authors } = props.data.allContentfulAuthor;